@@ -5,14 +5,12 @@ import android.graphics.Color;
 import android.graphics.Paint;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.JavascriptInterface;
 import android.widget.BaseAdapter;
 import android.widget.LinearLayout;
 import android.widget.TextView;
@@ -47,9 +45,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import wendu.dsbridge.CompletionHandler;
-import wendu.dsbridge.DWebView;
-
 import static com.android.jidian.client.bean.MainAppEventBean.PAYSUCCESSCLOSEORDER;
 import static com.android.jidian.client.bean.MainAppEventBean.PAYSUCCESSCLOSESHOP;
 
@@ -61,8 +56,7 @@ public class Newshoppay extends BaseActivity {
     TextView price_1, tv_title, total_price, cheap_price, subtotalPrice, select_coupons, tv_sure_order_package;
     @ViewById
     MyListView order_list;
-    @ViewById
-    DWebView web_view;
+
     List<Map<String, String>> datalist_order = new ArrayList<>();
     private String id = "";
     private String order_num = "";
@@ -96,12 +90,12 @@ public class Newshoppay extends BaseActivity {
             @Override
             public void handleMessage(Message msg) {
                 super.handleMessage(msg);
-                if (msg.what == 1) {
-                    web_view.setVisibility(View.VISIBLE);
-                    web_view.loadUrl(PubFunction.h5 + "Activity/layerFullCut.html");
-                } else {
-                    web_view.setVisibility(View.GONE);
-                }
+//                if (msg.what == 1) {
+//                    web_view.setVisibility(View.VISIBLE);
+//                    web_view.loadUrl(PubFunction.h5 + "Activity/layerFullCut.html");
+//                } else {
+//                    web_view.setVisibility(View.GONE);
+//                }
             }
         };
     }
@@ -121,12 +115,12 @@ public class Newshoppay extends BaseActivity {
         } else {
             HttpGetType(id);
         }
-        // 设置背景色
-        web_view.setBackgroundColor(0);
-        // 设置填充透明度
-        web_view.getBackground().setAlpha(0);
-        web_view.addJavascriptObject(new NewShopPayJsApi(), null);
-        web_view.getSettings().setTextZoom(100);
+//        // 设置背景色
+//        web_view.setBackgroundColor(0);
+//        // 设置填充透明度
+//        web_view.getBackground().setAlpha(0);
+//        web_view.addJavascriptObject(new NewShopPayJsApi(), null);
+//        web_view.getSettings().setTextZoom(100);
 //        web_view.loadUrl("https://testh5.halouhuandian.com/Activity/layerFullCut.html");
 //        tv_title.setOnClickListener(new View.OnClickListener() {
 //            @Override
@@ -394,7 +388,7 @@ public class Newshoppay extends BaseActivity {
                             }
                         }
                     }
-                    Newshoppay.MyBaseAdapter1 simpleAdapter = new MyBaseAdapter1(datalist_order, new int[]{R.layout.new_main_shop_order_item_top_v2, R.layout.main_shop_order_item_top2_v2, R.layout.main_shop_order_item_top_v23});
+                    Newshoppay.MyBaseAdapter1 simpleAdapter = new MyBaseAdapter1(datalist_order, new int[]{R.layout.u6_activity_pay_create_order_item, R.layout.u6_activity_pay_create_order_item_v2, R.layout.u6_activity_pay_create_order_item_v3});
                     order_list.setAdapter(simpleAdapter);
                     order_list.setDividerHeight(0);
                     simpleAdapter.notifyDataSetChanged();
@@ -513,22 +507,5 @@ public class Newshoppay extends BaseActivity {
         }
     }
 
-    public class NewShopPayJsApi {
-        //异步API
-        @JavascriptInterface
-        public void getInitInfo(Object msg, CompletionHandler<String> handler) {
-            Log.d("xiaoming1749", "getInitInfo: " + mFullCutData);
-            handler.complete(mFullCutData);
-        }
 
-        //异步API
-        @JavascriptInterface
-        public void OnClickExit(Object msg, CompletionHandler<String> handler) {
-            Log.d("xiaoming1749", "OnClickExit: ");
-//            web_view.setVisibility(View.GONE);
-            Message message = new Message();
-            message.what = 2;
-            mFullCutHandler.sendMessage(message);
-        }
-    }
 }
