@@ -2,29 +2,36 @@ package com.android.jidian.client.base;
 
 import android.app.Activity;
 import android.content.SharedPreferences;
-import android.content.res.Resources;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentActivity;
+
 import android.view.MotionEvent;
 
-import com.android.jidian.client.R;
+import com.android.jidian.client.base.broadcastManage.BroadcastManager;
 import com.android.jidian.client.base.inputManager.InputManager;
-import com.android.jidian.client.widgets.ProgressDialog;
-import com.gyf.barlibrary.ImmersionBar;
+import com.android.jidian.client.mvp.ui.dialog.DialogByLoading;
 
-import me.jessyan.autosize.AutoSizeConfig;
+import butterknife.ButterKnife;
+import butterknife.Unbinder;
 
 public class U6BaseActivity extends AppCompatActivity {
 
     protected static final String TAG = "BaseActivity";
-    protected Activity activity;
-
+    //上下文
+    protected FragmentActivity activity;
     //软键盘
     protected InputManager inputManager;
-
-    protected ProgressDialog progressDialog;
+    //butterKnife
+    protected Unbinder unbinder;
+    //appToken
+    protected String apptoken;
+    //广播
+    protected BroadcastManager broadcastManager;
+    //提示框
+    protected DialogByLoading progressDialog;
+    //用户数据
     protected SharedPreferences sharedPreferences;
     protected String uid;
 
@@ -32,18 +39,30 @@ public class U6BaseActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         activity = this;
+
         sharedPreferences = getSharedPreferences("userInfo", Activity.MODE_PRIVATE);
         uid = sharedPreferences.getString("id", "");
-        progressDialog = new ProgressDialog(this);
+        apptoken = sharedPreferences.getString("apptoken", "");
 
+        progressDialog = new DialogByLoading(this);
+
+        unbinder = ButterKnife.bind(this);
+
+        //初始化广播
+        registerFinishReceiver();
+        //初始化键盘
         inputInit();
+    }
+
+    //广播注册
+    protected void registerFinishReceiver() {
+        broadcastManager = new BroadcastManager(activity);
     }
 
     //键盘初始化
     private void inputInit(){
         inputManager = new InputManager(activity);
     }
-
     @Override
     public boolean dispatchTouchEvent(MotionEvent ev) {
         if (ev.getAction() == MotionEvent.ACTION_DOWN) {
@@ -57,10 +76,18 @@ public class U6BaseActivity extends AppCompatActivity {
         return onTouchEvent(ev);
     }
 
-
     @Override
     protected void onDestroy() {
-        inputManager.onDestroy();
         super.onDestroy();
+        if (unbinder != null) {
+            unbinder.unbind();
+        }
+        if(inputManager!=null){
+            inputManager.onDestroy();
+        }
+        if(inputManager!=null){
+            broadcastManager.onDestroy();
+        }
+
     }
 }
