@@ -27,8 +27,8 @@ public class IndexActivity extends U6BaseActivity {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
         setContentView(R.layout.u6_activity_index);
+        super.onCreate(savedInstanceState);
         ButterKnife.bind(this);
         init();
     }
